feat: refactor favorite button and recently viewed buttons
